% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{StudyLogistics}
\alias{StudyLogistics}
\title{Logistical Planning for Cross-Sectional Study}
\usage{
StudyLogistics(
  n_final,
  rejection_rate,
  eligibility_rate,
  participants_per_day,
  working_days_month = 22,
  language = "en"
)
}
\arguments{
\item{n_final}{Total sample required}

\item{rejection_rate}{Proportion of participants who will reject}

\item{eligibility_rate}{Proportion of eligible people}

\item{participants_per_day}{Participants that can be recruited per day}

\item{working_days_month}{Working days per month}

\item{language}{Language for messages: 'en' (English) or 'es' (Spanish). Default is 'en'}
}
\value{
List with logistics
}
\description{
Logistical Planning for Cross-Sectional Study
}
