% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{VerifyEPV}
\alias{VerifyEPV}
\title{Post-Study EPV Verification for Cross-Sectional Studies}
\usage{
VerifyEPV(n_final, n_with_outcome, k, language = "en")
}
\arguments{
\item{n_final}{Final sample obtained}

\item{n_with_outcome}{Participants with the observed outcome}

\item{k}{Number of variables in the model}

\item{language}{Language for messages: 'en' (English) or 'es' (Spanish). Default is 'en'}
}
\value{
List with observed EPV
}
\description{
Post-Study EPV Verification for Cross-Sectional Studies
}
